+++ src/main/java/org/opentripplanner/gtfs/mapping/ServiceDateMapper.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.gtfs.mapping;
 
 import org.opentripplanner.model.calendar.ServiceDate;
+import org.opentripplanner.model.calendar.ServiceDateInterval;
 
 /** Responsible for mapping GTFS ServiceDate into the OTP model. */
 class ServiceDateMapper {
@@ -10,4 +11,15 @@ class ServiceDateMapper {
                 null :
                 new ServiceDate(orginal.getYear(), orginal.getMonth(), orginal.getDay());
     }
+
+    static ServiceDateInterval mapServiceDateInterval(
+            org.onebusaway.gtfs.model.calendar.ServiceDate start,
+            org.onebusaway.gtfs.model.calendar.ServiceDate end
+    ) {
+        return new ServiceDateInterval(
+                mapServiceDate(start),
+                mapServiceDate(end)
+        );
+    }
+
 }