+++ src/main/java/org/opentripplanner/model/impl/EntityById.java
@@ -8,6 +8,8 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 /**
  * The purpose of this class is to provide a map from id to the corresponding entity.
@@ -43,13 +45,10 @@ public class EntityById<I extends Serializable, E extends TransitEntity<I>> {
     }
 
     /**
-     * Return a copy of the internal map. Changes in the source are not reflected
-     * in the destination (returned Map), and visa versa.
-     * <p>
-     * The returned map is immutable.
+     * Returns the number of key-value mappings in this map.
      */
-    Map<I, E> asImmutableMap() {
-        return Collections.unmodifiableMap(new HashMap<>(map));
+    public int size() {
+        return map.size();
     }
 
     public boolean containsKey(I id) {
@@ -59,6 +58,32 @@ public class EntityById<I extends Serializable, E extends TransitEntity<I>> {
     @Override
     public String toString() {
         return map.toString();
+    }
+
+    int removeIf(Predicate<E> test) {
+        Collection<E> newSet = map
+                .values()
+                .stream()
+                .filter(Predicate.not(test))
+                .collect(Collectors.toList());
+
+        int size = map.size();
+        if(newSet.size() == size) {
+            return 0;
+        }
+        map.clear();
+        addAll(newSet);
+        return size - map.size();
+    }
+
+    /**
+     * Return a copy of the internal map. Changes in the source are not reflected
+     * in the destination (returned Map), and visa versa.
+     * <p>
+     * The returned map is immutable.
+     */
+    Map<I, E> asImmutableMap() {
+        return Collections.unmodifiableMap(new HashMap<>(map));
     }
 
     /**