+++ src/test/java/org/opentripplanner/graph_builder/module/GtfsGraphBuilderModuleTest.java
@@ -9,6 +9,7 @@ import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.TransitEntity;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
+import org.opentripplanner.model.calendar.ServiceDateInterval;
 import org.opentripplanner.routing.graph.Graph;
 
 import java.io.IOException;
@@ -34,7 +35,7 @@ public class GtfsGraphBuilderModuleTest {
 
         List<GtfsBundle> bundleList = getGtfsAsBundleList(gtfs);
         bundleList.get(0).setDefaultBikesAllowed(false);
-        builder = new GtfsModule(bundleList);
+        builder = new GtfsModule(bundleList, ServiceDateInterval.unbounded());
 
         Graph graph = new Graph();
         builder.buildGraph(graph, _extra);
@@ -63,7 +64,7 @@ public class GtfsGraphBuilderModuleTest {
 
         List<GtfsBundle> bundleList = getGtfsAsBundleList(gtfs);
         bundleList.get(0).setDefaultBikesAllowed(true);
-        builder = new GtfsModule(bundleList);
+        builder = new GtfsModule(bundleList, ServiceDateInterval.unbounded());
 
         Graph graph = new Graph();
         builder.buildGraph(graph, _extra);