+++ docs/Changelog.md
@@ -18,6 +18,7 @@
 - Refactor GenericLocation/AStar/RoutingContext to allow multiple start vertices (#2887) 
 - New Transit search algorithm, Raptor, replaces the AStar for all transit searches. 
 - Ability to switch off the fare service(#2912).
+- Limit the transit service period(#2925).
 
 ## Ported over from the 1.x
 - Add application/x-protobuf to accepted protobuf content-types (#2839)